+++ test/pow_assent/strategies/discord_test.exs
@@ -1,7 +1,6 @@
 defmodule PowAssent.Strategy.DiscordTest do
-  use PowAssent.Test.Phoenix.ConnCase
+  use PowAssent.Test.OAuth2TestCase
 
-  import PowAssent.OAuthHelpers
   alias PowAssent.Strategy.Discord
 
   @user_response %{
@@ -12,34 +11,23 @@ defmodule PowAssent.Strategy.DiscordTest do
     "verified" => true,
     "email" => "nelly@discordapp.com"
   }
+  @user %{
+    "email" => "nelly@discordapp.com",
+    "name" => "Nelly",
+    "uid" => "80351110224678912",
+    "image" => "https://cdn.discordapp.com/avatars/80351110224678912/8342729096ea3675442027381ff50dfe"
+  }
 
-  setup :setup_bypass
-
-  test "authorize_url/2", %{conn: conn, config: config} do
-    assert {:ok, %{conn: _conn, url: url}} = Discord.authorize_url(config, conn)
+  test "authorize_url/2", %{config: config} do
+    assert {:ok, %{url: url}} = Discord.authorize_url(config)
     assert url =~ "/oauth2/authorize?client_id="
   end
 
-  describe "callback/2" do
-    setup %{conn: conn, config: config, bypass: bypass} do
-      params = %{"code" => "test", "redirect_uri" => "test"}
+  test "callback/2", %{config: config, callback_params: params, bypass: bypass} do
+    expect_oauth2_access_token_request(bypass, uri: "/oauth2/token")
+    expect_oauth2_user_request(bypass, @user_response, uri: "/users/@me")
 
-      {:ok, conn: conn, config: config, params: params, bypass: bypass}
-    end
-
-    test "normalizes data", %{conn: conn, config: config, params: params, bypass: bypass} do
-      expect_oauth2_access_token_request(bypass, uri: "/oauth2/token")
-      expect_oauth2_user_request(bypass, @user_response, uri: "/users/@me")
-
-      expected = %{
-        "email" => "nelly@discordapp.com",
-        "name" => "Nelly",
-        "uid" => "80351110224678912",
-        "image" => "https://cdn.discordapp.com/avatars/80351110224678912/8342729096ea3675442027381ff50dfe"
-      }
-
-      {:ok, %{user: user}} = Discord.callback(config, conn, params)
-      assert expected == user
-    end
+    assert {:ok, %{user: user}} = Discord.callback(config, params)
+    assert user == @user
   end
 end