+++ test/pow_assent/strategies/instagram_test.exs
@@ -1,7 +1,6 @@
 defmodule PowAssent.Strategy.InstagramTest do
-  use PowAssent.Test.Phoenix.ConnCase
+  use PowAssent.Test.OAuth2TestCase
 
-  import PowAssent.OAuthHelpers
   alias PowAssent.Strategy.Instagram
 
   @user_response %{
@@ -10,39 +9,30 @@ defmodule PowAssent.Strategy.InstagramTest do
     "full_name" => "Snoop Dogg",
     "profile_picture" => "..."
   }
+  @user %{
+    "image" => "...",
+    "name" => "Snoop Dogg",
+    "nickname" => "snoopdogg",
+    "uid" => "1574083"
+  }
 
-  setup :setup_bypass
-
-  test "authorize_url/2", %{conn: conn, config: config} do
-    assert {:ok, %{conn: _conn, url: url}} = Instagram.authorize_url(config, conn)
+  test "authorize_url/2", %{config: config} do
+    assert {:ok, %{url: url}} = Instagram.authorize_url(config)
     assert url =~ "/oauth/authorize?client_id="
   end
 
   describe "callback/2" do
-    setup %{conn: conn, config: config, bypass: bypass} do
-      params = %{"code" => "test", "redirect_uri" => "test"}
-
-      {:ok, conn: conn, config: config, params: params, bypass: bypass}
-    end
-
-    test "normalizes data", %{conn: conn, config: config, params: params, bypass: bypass} do
+    test "normalizes data", %{config: config, callback_params: params, bypass: bypass} do
       expect_oauth2_access_token_request(bypass, uri: "/oauth/token", params: %{access_token: "access_token", user: @user_response})
 
-      expected = %{
-        "image" => "...",
-        "name" => "Snoop Dogg",
-        "nickname" => "snoopdogg",
-        "uid" => "1574083"
-      }
-
-      {:ok, %{user: user}} = Instagram.callback(config, conn, params)
-      assert expected == user
+      assert {:ok, %{user: user}} = Instagram.callback(config, params)
+      assert user == @user
     end
 
-    test "handles error", %{config: config, conn: conn, params: params} do
-      config = Keyword.put(config, :site, "http://localhost:8888")
+    test "handles error", %{config: config, callback_params: params, bypass: bypass} do
+      Bypass.down(bypass)
 
-      assert {:error, %{conn: %Plug.Conn{}, error: %PowAssent.RequestError{error: :unreachable}}} = Instagram.callback(config, conn, params)
+      assert {:error, %PowAssent.RequestError{error: :unreachable}} = Instagram.callback(config, params)
     end
   end
 end