+++ test/pow_assent/strategies/oauth_test.exs
@@ -1,68 +1,67 @@
 defmodule PowAssent.Strategy.OAuthTest do
-  use PowAssent.Test.Phoenix.ConnCase
+  use PowAssent.Test.OAuthTestCase
 
-  import PowAssent.OAuthHelpers
   alias PowAssent.{RequestError, Strategy.OAuth}
 
-  setup context do
-    setup_bypass_oauth(context, user_url: "/api/user")
-  end
-
   describe "authorize_url/2" do
-    test "returns url", %{conn: conn, config: config, bypass: bypass} do
+    test "returns url", %{config: config, bypass: bypass} do
       expect_oauth_request_token_request(bypass)
 
-      assert {:ok, %{conn: _conn, url: url}} = OAuth.authorize_url(config, conn)
-      assert url =~ bypass_server(bypass) <> "/oauth/authenticate?oauth_token=token"
+      assert {:ok, %{url: url}} = OAuth.authorize_url(config)
+      assert url =~ "http://localhost:#{bypass.port}/oauth/authenticate?oauth_token=token"
     end
 
-    test "falls back to parse response as uri query", %{conn: conn, config: config, bypass: bypass} do
+    test "parses URI query response", %{config: config, bypass: bypass} do
       expect_oauth_request_token_request(bypass, content_type: "text/html", params: URI.encode_query(%{oauth_token: "token", oauth_token_secret: "token_secret"}))
 
-      assert {:ok, _resp} = OAuth.authorize_url(config, conn)
+      assert {:ok, %{url: url}} = OAuth.authorize_url(config)
+      assert url =~ "http://localhost:#{bypass.port}/oauth/authenticate?oauth_token=token"
     end
 
-    test "bubbles up unexpected response", %{conn: conn, config: config, bypass: bypass} do
+    test "bubbles up unexpected response with HTTP status 200", %{config: config, bypass: bypass} do
       expect_oauth_request_token_request(bypass, params: %{"error_code" => 215, "error_message" => "Bad Authentication data."})
 
-      assert {:error, %{conn: _conn, error: %RequestError{error: :unexpected_response}}} = OAuth.authorize_url(config, conn)
+      assert {:error, %RequestError{error: :unexpected_response}} =OAuth.authorize_url(config)
     end
 
-    test "bubbles up error response", %{conn: conn, config: config, bypass: bypass} do
+    test "bubbles up error response", %{config: config, bypass: bypass} do
       expect_oauth_request_token_request(bypass, status_code: 500, params: %{"error_code" => 215, "error_message" => "Bad Authentication data."})
 
-      assert {:error, %{conn: _conn, error: %RequestError{error: :invalid_server_response}}} = OAuth.authorize_url(config, conn)
+      assert {:error, %RequestError{error: :invalid_server_response}} = OAuth.authorize_url(config)
     end
 
-    test "bubbles up json error response", %{conn: conn, config: config, bypass: bypass} do
+    test "bubbles up json error response", %{config: config, bypass: bypass} do
       expect_oauth_request_token_request(bypass, status_code: 500, content_type: "application/json", params: %{"errors" => [%{"code" => 215, "message" => "Bad Authentication data."}]})
 
-      assert {:error, %{conn: _conn, error: %RequestError{error: :invalid_server_response}}} = OAuth.authorize_url(config, conn)
+      assert {:error, %RequestError{error: :invalid_server_response}} = OAuth.authorize_url(config)
     end
 
-    test "bubbles up network error", %{conn: conn, config: config, bypass: bypass} do
+    test "bubbles up network error", %{config: config, bypass: bypass} do
       Bypass.down(bypass)
 
-      assert {:error, %{conn: _conn, error: %PowAssent.RequestError{error: :unreachable}}} = OAuth.authorize_url(config, conn)
+      assert {:error, %PowAssent.RequestError{error: :unreachable}} = OAuth.authorize_url(config)
     end
   end
 
   describe "callback/2" do
-    test "normalizes data", %{conn: conn, config: config, params: params, bypass: bypass} do
+    setup %{config: config} = context do
+      config = Keyword.put(config, :user_url, "/api/user")
+
+      %{context | config: config}
+    end
+
+    test "normalizes data", %{config: config, callback_params: params, bypass: bypass} do
       expect_oauth_access_token_request(bypass)
       expect_oauth_user_request(bypass, %{email: nil})
 
-      expected = %{"email" => nil}
-
-      {:ok, %{user: user}} = OAuth.callback(config, conn, params)
-      assert expected == user
+      assert {:ok, %{user: %{"email" => nil}}} = OAuth.callback(config, params)
     end
 
-    test "bubbles up error response", %{conn: conn, config: config, params: params, bypass: bypass} do
+    test "bubbles up error response", %{config: config, callback_params: params, bypass: bypass} do
       expect_oauth_access_token_request(bypass)
       expect_oauth_user_request(bypass, %{error: "Unknown error"}, status_code: 500)
 
-      {:error, %{conn: _conn, error: %RequestError{error: :invalid_server_response}}} = OAuth.callback(config, conn, params)
+      assert {:error, %RequestError{error: :invalid_server_response}} = OAuth.callback(config, params)
     end
   end
 end